requirements.txt
# requirements.txt
ollama-python>=0.1.0
flask==2.2.3
numpy==1.24.2
PyPDF2==3.0.1
pandas==2.0.3
python-docx==0.8.11